Moved formatting functions together
